requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers==0.11.1
transformers==4.25.1
huggingface_hub==0.16.4
imageio==2.27.0
imageio-ffmpeg==0.4.9
decord==0.6.0
omegaconf==2.3.0
gradio==3.36.1
safetensors
einops
numpy
pillow
tqdm
flask
flask-cors
flask-socketio
opencv-python
bitsandbytes
accelerate<0.21.0